requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
PyJWT==2.10.1
python-multipart==0.0.20
psycopg2-binary==2.9.10
python-dotenv==1.1.1
